requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

# API & Testing
requests==2.31.0
httpx==0.25.1
pytest==7.4.3

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0

# Database
sqlalchemy==2.0.23

# Data Processing
python-slugify==8.0.1

# LLM & AI
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-groq>=0.2.0,<0.3.0
langchain-tavily>=0.1.0
langgraph>=0.2.0
google-generativeai>=0.8.0
tavily-python>=0.1.0
tiktoken>=0.5.0

# UI (optional for production)
streamlit>=1.28.0

# Server & Monitoring
gunicorn>=21.2.0
watchdog>=3.0.0